+++ Editor/Unity.Platforms.Windows.Build/NonIncremental/Steps/WindowsProduceArtifactStep.cs
@@ -1,0 +1,23 @@
+using System.IO;
+using Unity.Build;
+using BuildReport = UnityEditor.Build.Reporting.BuildReport;
+
+namespace Unity.Platforms.Windows.Build
+{
+    [BuildStep(Description = "Producing Windows Artifacts")]
+    sealed class WindowsProduceArtifactStep : BuildStepBase
+    {
+        public override BuildResult Run(BuildContext context)
+        {
+            var report = context.GetValue<BuildReport>();
+            if (report == null)
+            {
+                return context.Failure($"Could not retrieve {nameof(BuildReport)} from build context.");
+            }
+
+            var artifact = context.GetOrCreateValue<WindowsArtifact>();
+            artifact.OutputTargetFile = new FileInfo(report.summary.outputPath);
+            return context.Success();
+        }
+    }
+}