+++ Editor/Unity.Platforms.Windows.Build/NonIncremental/WindowsArtifact.cs
@@ -1,0 +1,10 @@
+using System.IO;
+using Unity.Build;
+
+namespace Unity.Platforms.Windows.Build
+{
+    sealed class WindowsArtifact : IBuildArtifact
+    {
+        public FileInfo OutputTargetFile;
+    }
+}