+++ Editor/Unity.Platforms.Windows.Build/WindowsMenuItem.cs
@@ -2,16 +2,14 @@
 using Unity.Build.Classic;
 using Unity.Build.Common;
 using Unity.Build.Editor;
+using Unity.BuildSystem.NativeProgramSupport;
 using UnityEditor;
-using UnityEngine;
-using BuildPipeline = Unity.Build.BuildPipeline;
 
 namespace Unity.Platforms.Windows.Build
 {
-    public static class MenuItemWindows
+    static class WindowsMenuItem
     {
         const string k_CreateBuildConfigurationAssetClassic = BuildConfigurationMenuItem.k_BuildConfigurationMenu + "Windows Classic Build Configuration";
-        const string k_BuildPipelineClassicAssetPath = "Packages/com.unity.platforms.windows/Editor/Unity.Platforms.Windows.Build/Assets/Windows Classic.buildpipeline";
 
         [MenuItem(k_CreateBuildConfigurationAssetClassic, true)]
         static bool CreateBuildConfigurationAssetClassicValidation()
@@ -22,15 +20,13 @@ namespace Unity.Platforms.Windows.Build
         [MenuItem(k_CreateBuildConfigurationAssetClassic)]
         static void CreateBuildConfigurationAssetClassic()
         {
-            var pipeline = AssetDatabase.LoadAssetAtPath<BuildPipeline>(k_BuildPipelineClassicAssetPath);
             Selection.activeObject = BuildConfigurationMenuItem.CreateAssetInActiveDirectory(
-                "WindowsClassic", new GeneralSettings(), new SceneList(), new ClassicBuildProfile
+                "WindowsClassic",
+                new GeneralSettings(),
+                new SceneList(),
+                new ClassicBuildProfile
                 {
-#if UNITY_2020_1_OR_NEWER
-                    Pipeline = new LazyLoadReference<BuildPipeline> { asset = pipeline }
-#else
-                    Pipeline = pipeline
-#endif
+                    Platform = new WindowsPlatform()
                 });
         }
     }